+++ source/exhumed/src/sequence.h
@@ -142,11 +142,11 @@ int seq_GetSeqPicnum2(short nSeq, short nFrame);
 int seq_GetSeqPicnum(short nSeq, short edx, short ebx);
 void seq_DrawStatusSequence(short nSequence, uint16_t edx, short ebx);
 
-int seq_DrawGunSequence(int nSeqOffset, short dx, int xOffs, int yOffs, int nShade, int nPal);
+int seq_DrawGunSequence(int nSeqOffset, short dx, double xOffs, double yOffs, int nShade, int nPal);
 short seq_GetFrameFlag(short val, short nFrame);
 int seq_PlotSequence(short nSprite, short edx, short nFrame, short ecx);
 int seq_PlotArrowSequence(short nSprite, short nSeq, int nVal);
-void seq_DrawPilotLightSeq(int xOffset, int yOffset);
+void seq_DrawPilotLightSeq(double xOffset, double yOffset);
 
 END_PS_NS
